+++ lib/widgets/dialog_full_screen.dart
@@ -31,9 +31,25 @@ class AutoSizeDialog extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    ShapeBorder? dialogShape;
+    EdgeInsets? insetPadding;
+
+    if (isSmallDevice(context)) {
+      insetPadding = EdgeInsets.zero;
+    } else {
+      dialogShape = RoundedRectangleBorder(
+        borderRadius: const BorderRadius.all(Radius.circular(8)),
+        side: BorderSide(
+          width: 2.0, // Adjust border width as needed
+          color: Theme.of(context).dividerColor, // Set your desired border color here
+        ),
+      );
+      insetPadding = const EdgeInsets.symmetric(horizontal: 40.0, vertical: 24.0);
+    }
+
     return Dialog(
-      insetPadding:
-          isSmallDevice(context) ? EdgeInsets.zero : const EdgeInsets.symmetric(horizontal: 40.0, vertical: 24.0),
+      shape: dialogShape,
+      insetPadding: insetPadding,
       // Set elevation to 0 to remove default shadow
       elevation: 0.0,
       child: ConstrainedBox(